Made empty database provider test only
