+++ go/libraries/doltcore/sqle/dsess/revision_db_provider.go
@@ -86,56 +86,3 @@ type DoltDatabaseProvider interface {
 	// database.
 	SessionDatabase(ctx *sql.Context, dbName string) (SessionDatabase, bool, error)
 }
-
-// EmptyDatabaseProvider returns a DoltDatabaseProvider with no content. Only suitable for test cases 
-func EmptyDatabaseProvider() DoltDatabaseProvider {
-	return emptyRevisionDatabaseProvider{}
-}
-
-type emptyRevisionDatabaseProvider struct {
-	sql.DatabaseProvider
-}
-
-func (e emptyRevisionDatabaseProvider) DbState(ctx *sql.Context, dbName string, defaultBranch string) (InitialDbState, error) {
-	return InitialDbState{}, sql.ErrDatabaseNotFound.New(dbName)
-}
-
-func (e emptyRevisionDatabaseProvider) DropDatabase(ctx *sql.Context, name string) error {
-	return nil
-}
-
-func (e emptyRevisionDatabaseProvider) GetRevisionForRevisionDatabase(_ *sql.Context, _ string) (string, string, error) {
-	return "", "", nil
-}
-
-func (e emptyRevisionDatabaseProvider) IsRevisionDatabase(_ *sql.Context, _ string) (bool, error) {
-	return false, nil
-}
-
-func (e emptyRevisionDatabaseProvider) GetRemoteDB(ctx context.Context, format *types.NomsBinFormat, r env.Remote, withCaching bool) (*doltdb.DoltDB, error) {
-	return nil, nil
-}
-
-func (e emptyRevisionDatabaseProvider) FileSystem() filesys.Filesys {
-	return nil
-}
-
-func (e emptyRevisionDatabaseProvider) FileSystemForDatabase(dbname string) (filesys.Filesys, error) {
-	return nil, nil
-}
-
-func (e emptyRevisionDatabaseProvider) CloneDatabaseFromRemote(ctx *sql.Context, dbName, branch, remoteName, remoteUrl string, remoteParams map[string]string) error {
-	return nil
-}
-
-func (e emptyRevisionDatabaseProvider) CreateDatabase(ctx *sql.Context, dbName string) error {
-	return nil
-}
-
-func (e emptyRevisionDatabaseProvider) RevisionDbState(_ *sql.Context, revDB string) (InitialDbState, error) {
-	return InitialDbState{}, sql.ErrDatabaseNotFound.New(revDB)
-}
-
-func (e emptyRevisionDatabaseProvider) SessionDatabase(ctx *sql.Context, dbName string) (SessionDatabase, bool, error) {
-	return nil, false, nil
-}
